chore: export ColumnarViewModel properly ht Marco
